requirements.txt
streamlit
ultralytics
opencv-python
pillow
